+++ CHANGELOG.md
@@ -7,6 +7,7 @@ just to make sure i was not sleeping all the time.
 ### 8th April, 2025
 - added more applications
 - ui/ux design applications section
+- added reset filters button
 
 ### 7th April, 2025
 - all applications were added